+++ dotnet/GraphQL/Mutation.cs
@@ -2,8 +2,11 @@
 using FlowFinance.Services;
 using GraphQL;
 using GraphQL.Types;
+using GraphQL.Upload.AspNetCore;
+using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Text;
 
 namespace FlowFinance.GraphQL
@@ -29,12 +32,17 @@ namespace FlowFinance.GraphQL
             Field<ApplicationResultType>(
                 "processApplication",
                 arguments: new QueryArguments(
-                    new QueryArgument<NonNullGraphType<ApplicationInputType>> { Name = "application", Description = "Application Input" }
+                    new QueryArgument<NonNullGraphType<ApplicationInputType>> { Name = "application", Description = "Application Input" },
+                    new QueryArgument<NonNullGraphType<UploadGraphType>> { Name = "businessInfoFile" },
+                    new QueryArgument<NonNullGraphType<UploadGraphType>> { Name = "personalInfoFile" }
                 ),
                 resolve: context =>
                 {
                     var applicationInput = context.GetArgument<ApplicationInput>("application");
-                    return flowFinancePaymentService.ProcessApplication(applicationInput);
+                    var businessInfoFile = context.GetArgument<IFormFile>("businessInfoFile");
+                    var personalInfoFile = context.GetArgument<IFormFile>("personalInfoFile");
+
+                    return flowFinancePaymentService.ProcessApplication(applicationInput, businessInfoFile, personalInfoFile);
                 });
         }
     }